+++ .gitignore
@@ -22,6 +22,9 @@ target/
 *.iml
 *.ipr
 
+### Visual Studio Code ###
+.vscode
+
 ### NetBeans ###
 /nbproject/private/
 /build/
