--- src/main/java/com/denachina/shadow/ShadowApplication.java
@@ -1,14 +1,0 @@
-package com.denachina.shadow;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ShadowApplication{
-
-	public static void main(String[] args) {
-		SpringApplication.run(ShadowApplication.class, args);
-	}
-
-}
-
