--- src/main/java/com/denachina/shadow/config/dbconfig/DBContextHolder.java
@@ -1,20 +1,0 @@
-package com.denachina.shadow.config.dbconfig;
-
-public class DBContextHolder {
-    public static final String R = "datasourceR";
-    public static final String W = "datasourceW";
-
-    private static final ThreadLocal<String> CONTEXT_HOLDER = new ThreadLocal<String>();
-
-    public static void setDBType(String dbType) {
-        CONTEXT_HOLDER.set(dbType);
-    }
-
-    public static String getDBType() {
-        return CONTEXT_HOLDER.get();
-    }
-
-    public static void clearDBType() {
-        CONTEXT_HOLDER.remove();
-    }
-}
