--- src/main/java/com/denachina/shadow/config/dbconfig/DynamicDataSource.java
@@ -1,10 +1,0 @@
-package com.denachina.shadow.config.dbconfig;
-
-import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
-
-public class DynamicDataSource extends AbstractRoutingDataSource {
-    @Override
-    protected Object determineCurrentLookupKey() {
-        return DBContextHolder.getDBType();
-    }
-}
