--- src/main/java/com/denachina/shadow/service/UserDataService.java
@@ -1,15 +1,0 @@
-package com.denachina.shadow.service;
-
-import java.util.List;
-
-import com.denachina.shadow.dao.UserData;
-
-public interface UserDataService {
-
-    List<UserData> getAllUserData();
-
-    UserData insertUserData(UserData userData);
-
-    int deleteUserData(Integer UserId);
-
-}
