+++ src/main/java/com/github/shadow/controller/SysUserController.java
@@ -1,0 +1,18 @@
+package com.github.shadow.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.stereotype.Controller;
+
+/**
+ * <p>
+ * 用户表 前端控制器
+ * </p>
+ *
+ * @author xuwen
+ * @since 2023-07-12
+ */
+@Controller
+@RequestMapping("/dao/sysUser")
+public class SysUserController {
+
+}
