+++ src/main/java/com/github/shadow/controller/UserDataController.java
@@ -1,0 +1,18 @@
+package com.github.shadow.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.stereotype.Controller;
+
+/**
+ * <p>
+ * 用户信息表 前端控制器
+ * </p>
+ *
+ * @author xuwen
+ * @since 2023-07-12
+ */
+@Controller
+@RequestMapping("/dao/userData")
+public class UserDataController {
+
+}
