+++ src/main/java/com/github/shadow/mapper/SysUserMapper.java
@@ -1,0 +1,16 @@
+package com.github.shadow.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.github.shadow.entity.SysUser;
+
+/**
+ * <p>
+ * 用户表 Mapper 接口
+ * </p>
+ *
+ * @author xuwen
+ * @since 2023-07-12
+ */
+public interface SysUserMapper extends BaseMapper<SysUser> {
+
+}
