+++ src/main/java/com/github/shadow/mapper/UserDataMapper.java
@@ -1,0 +1,16 @@
+package com.github.shadow.mapper;
+
+import com.github.shadow.entity.UserData;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+
+/**
+ * <p>
+ * 用户信息表 Mapper 接口
+ * </p>
+ *
+ * @author xuwen
+ * @since 2023-07-12
+ */
+public interface UserDataMapper extends BaseMapper<UserData> {
+
+}
