+++ src/main/java/com/github/shadow/service/IUserDataService.java
@@ -1,0 +1,16 @@
+package com.github.shadow.service;
+
+import com.github.shadow.entity.UserData;
+import com.baomidou.mybatisplus.extension.service.IService;
+
+/**
+ * <p>
+ * 用户信息表 服务类
+ * </p>
+ *
+ * @author xuwen
+ * @since 2023-07-12
+ */
+public interface IUserDataService extends IService<UserData> {
+
+}
