--- src/main/resources/db.properties
@@ -1,7 +1,0 @@
-mysql.datasource.diverClassName=com.mysql.cj.jdbc.Driver
-mysql.datasource.r.jdbc-url=jdbc:mysql://mysql:3306/shadow?userSSL=false
-mysql.datasource.r.username=${DB_USERNAME:root}
-mysql.datasource.r.password=${DB_PASSWORD:root}
-mysql.datasource.w.jdbc-url=jdbc:mysql://mysql:3306/shadow?userSSL=false
-mysql.datasource.w.username=${DB_USERNAME:root}
-mysql.datasource.w.password=${DB_PASSWORD:root}
