--- src/main/resources/hibernate.properties
@@ -1,24 +1,0 @@
-# Hibernate SQL Dialects (hibernate.dialect)
-#hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-
-# Naming strategies
-#hibernate.physical_naming_strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-
-# SpringPhysicalNamingStrategy
-hibernate.physical_naming_strategy=org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy
-
-
-# Automatic schema generation
-hibernate.hbm2ddl.auto=validate
-
-# Statement logging and statistics
-hibernate.show_sql=true
-#hibernate.format_sql=true
-
-hibernate.temp.use_jdbc_metadata_defaults=false
-
-#hibernate.generate_statistics=true
-
-hibernate.use_sql_comments=true
-
